pyproject.toml
[project]
name = "linear-updates"
version = "0.1.0"
description = "Draft weekly stakeholder updates from Linear cycles."
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
  "httpx>=0.27.0",
  "python-dotenv>=1.0.1",
  "rich>=13.7.1",
  "typer>=0.12.3",
]

[project.scripts]
linear-updates = "linear_updates.cli:app"

[dependency-groups]
dev = [
  "pytest>=8.0",
  "ruff>=0.8",
  "mypy>=1.10",
]

[build-system]
requires = ["hatchling>=1.24.2"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/linear_updates"]

[tool.ruff]
line-length = 100
target-version = "py311"
src = ["src"]

[tool.ruff.lint]
select = ["E", "F", "I", "UP", "B", "SIM"]
ignore = ["B008"]  # Allow function calls in defaults (required for Typer)

[tool.mypy]
python_version = "3.11"
warn_return_any = false
warn_unused_ignores = true
check_untyped_defs = true
disallow_untyped_defs = false  # Relaxed for existing code

[tool.pytest.ini_options]
testpaths = ["tests"]

